refactor(database): implement helper SerializeByValueType
